09.16.24 Updated Scripts
Experimenting with simpler system, snow falling. Still a natural phenomenon that has an evident direction. Will simplify down further, but am using this as an example to explore the concept and image prcoessing for a different type of video, and fine tune the pre-processing steps. Would love recommendations on any datasets that have a simple ball flying through the air.